Refactor of GEMINI
I have separated out the RTI-specific network code, providing stubs in the /net subfolder. The GUI elements have been separated with proxies that conform to the same interface. Jenn can start work on the interface and connect to the HTML site. Even so, there is still a problem with the interface as Qt doesn't allow an interface to have signals and slots. Also QObjects can't have two parent classes that are QObjects so there is a conflict. This creates problems in all the qtgui classes and the ProxySet class. All the Driver code in /net is a starter point for Kurt.